Разбивает main.js на модули
